requirements.txt
pandas==2.2.2
clickhouse-driver==0.2.9
pyspark==3.5.4
pymongo==4.10.0
psycopg2-binary==2.9.10
python-dotenv==0.21.0
kafka-python==2.2.12
